back/requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==42.0.8
curl_cffi==0.11.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.11
django-allauth==0.54.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.14.0
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
joblib==1.5.1
lxml==5.4.0
multitasking==0.0.11
numpy==2.0.2
oauthlib==3.2.2
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soupsieve==2.7
sqlparse==0.4.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.61
